Improved precision for waveform generator.

+++ dtmf.h
@@ -61,11 +61,11 @@ class DTMF {
   private:
     // The DTMF wave generator
     WAVE wave;
-    // Wave steps
+    // Wave steps (Q6.2)
     uint8_t stpRows[ROWSCOLS];
     uint8_t stpCols[ROWSCOLS];
-    // Wave indices for row and col frequencies
-    uint8_t rowIdx, colIdx;
+    // Wave indices for row and col frequencies (Q14.2)
+    uint16_t rowIdx, colIdx;
     // Identified row and col
     uint8_t row, col;
     // Wave generator status: DISABLED, WAVE, SILENCE
